Prevent cast warning. Part of bug #318095. Patch by Stanislav Brabec. Fix
2005-10-06  noel  <vnoel@cox.net>

	* log_repaint.c (logview_draw_log_lines): Prevent cast
	warning. Part of bug #318095. Patch by Stanislav Brabec.
	* about.c: Fix missing include. Part of bug #318095. Patch
	by Stanislav Brabec.

+++ logview/log_repaint.c
@@ -413,7 +413,7 @@ logview_draw_log_lines (LogviewWindow *window, Log *current_log)
                 utf8 = GetDateHeader (line);
                 gtk_tree_store_append (GTK_TREE_STORE (model), &iter, NULL);
                 gtk_tree_store_set (GTK_TREE_STORE (model), &iter,
-                                    DATE, utf8, -1);
+                                    DATE, utf8, (long) -1);
                 
                 /* store pointer to the date headers, using the month and day as the key */
